requirements.txt
absl-py==2.0.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
basicsr==1.4.2
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
docker-pycreds==0.4.0
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.9.2
future==0.18.3
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
huggingface-hub==0.18.0
humanfriendly==10.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.9.0
lmdb==1.4.1
lpips==0.1.4
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.15.1
opencv-python==4.8.1.78
packaging==23.2
pathtools==0.1.2
Pillow==10.1.0
platformdirs==3.11.0
proglog==0.1.10
protobuf==4.24.4
psutil==5.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scikit-image==0.21.0
scipy==1.10.1
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
sk-video==1.1.10
smmap==5.0.1
sympy==1.12
tb-nightly==2.14.0a20230808
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tifffile==2023.7.10
timm==0.9.7
tomli==2.0.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchmetrics==1.2.0
torchvision==0.13.1+cu116
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.6
wandb==0.15.12
Werkzeug==3.0.0
yapf==0.40.2
zipp==3.17.0
